backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pymongo
python-dotenv
imbalanced-learn
pytest
httpx
pytest-asyncio
